+++ ext/Server/Bot.lua
@@ -312,11 +312,9 @@ end
 
 ---@param p_ShootBackAfterHit boolean
 ---@param p_Player Player | nil
----@param p_EnemyVehicleType VehicleTypes
----@param p_BotVehicleType VehicleTypes
+---@param p_CheckShootTimer boolean
 ---@return boolean
----@return integer|nil
-function Bot:IsReadyToAttack(p_ShootBackAfterHit, p_Player, p_EnemyVehicleType, p_BotVehicleType)
+function Bot:IsReadyToAttack(p_ShootBackAfterHit, p_Player, p_CheckShootTimer)
 	-- update timers first
 	if self._ShootPlayerName == '' then
 		self._DoneShootDuration = 0.0
@@ -333,9 +331,8 @@ function Bot:IsReadyToAttack(p_ShootBackAfterHit, p_Player, p_EnemyVehicleType, 
 		return false
 	end
 
-	local s_NewAttackPriority = self:GetAttackPriority(p_EnemyVehicleType, p_BotVehicleType)
-	if s_NewAttackPriority > self.m_AttackPriority then
-		return true, s_NewAttackPriority
+	if not p_CheckShootTimer then
+		return true
 	end
 
 	if self._ShootPlayerName == '' or
@@ -343,16 +340,15 @@ function Bot:IsReadyToAttack(p_ShootBackAfterHit, p_Player, p_EnemyVehicleType, 
 		(self.m_InVehicle and (self._DoneShootDuration > Config.BotVehicleMinTimeShootAtPlayer)) or
 		(not self.m_InVehicle and (self._DoneShootDuration > Config.BotMinTimeShootAtPlayer)) or
 		(self.m_KnifeMode and self._ShootModeTimer > ((Config.BotMinTimeShootAtPlayer * 0.5))) then
-		return true, s_NewAttackPriority
+		return true
 	else
 		return false
 	end
 end
 
----@param p_EnemyVehicleType VehicleTypes
----@param p_BotVehicleType VehicleTypes
----@return integer
-function Bot:GetAttackPriority(p_EnemyVehicleType, p_BotVehicleType)
+function Bot:GetAttackPriority(p_EnemyVehicleType)
+	local s_BotVehicleType = m_Vehicles:VehicleType(self.m_ActiveVehicle)
+
 	if self.m_SecondaryGadget ~= nil then
 		if self.m_SecondaryGadget.type == WeaponTypes.MissileAir
 			and m_Vehicles:IsAirVehicleType(p_EnemyVehicleType)
@@ -365,7 +361,7 @@ function Bot:GetAttackPriority(p_EnemyVehicleType, p_BotVehicleType)
 		end
 	end
 
-	if m_Vehicles:IsAirVehicleType(p_BotVehicleType) then
+	if m_Vehicles:IsAirVehicleType(s_BotVehicleType) then
 		if m_Vehicles:IsAirVehicleType(p_EnemyVehicleType) then
 			return 3
 		elseif m_Vehicles:IsArmoredVehicleType(p_EnemyVehicleType) then
@@ -373,7 +369,7 @@ function Bot:GetAttackPriority(p_EnemyVehicleType, p_BotVehicleType)
 		end
 	end
 
-	if m_Vehicles:IsArmoredVehicleType(p_BotVehicleType)
+	if m_Vehicles:IsArmoredVehicleType(s_BotVehicleType)
 		and m_Vehicles:IsArmoredVehicleType(p_EnemyVehicleType)
 	then
 		return 2
@@ -468,14 +464,6 @@ end
 ---@param p_IgnoreYaw boolean
 ---@return boolean
 function Bot:ShootAt(p_Player, p_IgnoreYaw)
-	local s_EnemyVehicleType = m_Vehicles:FindOutVehicleType(p_Player)
-	local s_BotVehicleType = m_Vehicles:VehicleType(self.m_ActiveVehicle)
-
-	local s_IsReady, s_NewPriority = self:IsReadyToAttack(p_IgnoreYaw, p_Player, s_EnemyVehicleType, s_BotVehicleType)
-	if not s_IsReady or self._Shoot == false then
-		return false
-	end
-
 	if p_IgnoreYaw and self._DefendTimer == 0.0 then -- was hit, not in defend-mode, check for special behavior
 		if self.m_Behavior == BotBehavior.DontShootBackHide then
 			self._ActionTimer = 7.0
@@ -498,8 +486,18 @@ function Bot:ShootAt(p_Player, p_IgnoreYaw)
 	end
 
 	-- Check for vehicles.
+	local s_Type = m_Vehicles:FindOutVehicleType(p_Player)
+
 	-- Don't shoot at stationary AA.
-	if s_EnemyVehicleType == VehicleTypes.StationaryAA then
+	if s_Type == VehicleTypes.StationaryAA then
+		return false
+	end
+
+	local s_NewAttackPriority = self:GetAttackPriority(s_Type)
+
+	if not (s_NewAttackPriority > self.m_AttackPriority or self:IsReadyToAttack(p_IgnoreYaw, p_Player, true))
+		or self._Shoot == false
+	then
 		return false
 	end
 
@@ -507,7 +505,7 @@ function Bot:ShootAt(p_Player, p_IgnoreYaw)
 	self._DistanceToPlayer = 0.0
 	local s_PlayerPos = nil
 	local s_TargetPos = nil
-	if s_EnemyVehicleType == VehicleTypes.MavBot or s_EnemyVehicleType == VehicleTypes.MobileArtillery then
+	if s_Type == VehicleTypes.MavBot or s_Type == VehicleTypes.MobileArtillery then
 		s_TargetPos = p_Player.controlledControllable.transform.trans:Clone()
 	else
 		s_TargetPos = p_Player.soldier.worldTransform.trans:Clone()
@@ -526,8 +524,8 @@ function Bot:ShootAt(p_Player, p_IgnoreYaw)
 		s_IsSniper = true
 	end
 	local s_VehicleAttackMode = nil
-	if s_EnemyVehicleType ~= VehicleTypes.NoVehicle then
-		s_VehicleAttackMode = m_Vehicles:CheckForVehicleAttack(s_EnemyVehicleType, self._DistanceToPlayer, self.m_SecondaryGadget,
+	if s_Type ~= VehicleTypes.NoVehicle then
+		s_VehicleAttackMode = m_Vehicles:CheckForVehicleAttack(s_Type, self._DistanceToPlayer, self.m_SecondaryGadget,
 			self.m_InVehicle, s_IsSniper)
 		if s_VehicleAttackMode == VehicleAttackModes.NoAttack then
 			return false
@@ -541,19 +539,19 @@ function Bot:ShootAt(p_Player, p_IgnoreYaw)
 		return false
 	end
 
-	self._ShootPlayerVehicleType = s_EnemyVehicleType
+	self._ShootPlayerVehicleType = s_Type
 
 	local s_DifferenceYaw = 0
 	local s_Pitch = 0
 	local s_FovHalf = 0
 	local s_PitchHalf = 0
 
-	if self.m_InVehicle and m_Vehicles:IsVehicleType(self.m_ActiveVehicle, VehicleTypes.LightAA) and not m_Vehicles:IsAirVehicleType(s_EnemyVehicleType) then
+	if self.m_InVehicle and m_Vehicles:IsVehicleType(self.m_ActiveVehicle, VehicleTypes.LightAA) and not m_Vehicles:IsAirVehicleType(s_Type) then
 		return false
 	end
 
 	-- If target is air-vehicle and bot is in AA → ignore yaw.
-	if m_Vehicles:IsAirVehicleType(s_EnemyVehicleType) then
+	if m_Vehicles:IsAirVehicleType(s_Type) then
 		if (self.m_InVehicle and m_Vehicles:IsAAVehicle(self.m_ActiveVehicle)) or
 			(s_VehicleAttackMode == VehicleAttackModes.AttackWithMissileAir) then
 			p_IgnoreYaw = true
@@ -630,8 +628,8 @@ function Bot:ShootAt(p_Player, p_IgnoreYaw)
 				-- check for changed weapon
 				BotSpawner:UpdateGmWeapon(self)
 			end
-			self.m_AttackPriority = s_NewPriority
 			self._KillYourselfTimer = 0.0
+			self.m_AttackPriority = s_NewAttackPriority
 			return true
 		else
 			self._ShootPlayerName = ''