+++ ext/Server/BotManager.lua
@@ -87,7 +87,10 @@ function BotManager:OnUpdateManagerUpdate(p_DeltaTime, p_UpdatePass)
 		self._BotAttackBotTimer = self._BotAttackBotTimer + p_DeltaTime
 	end
 
-	if Config.BotsReviveBots and self._InitDone then
+	if Config.BotsReviveBots
+		and self._InitDone
+		and not Globals.IsGm
+	then
 		if self._BotReviveBotTimer >= Registry.GAME_RAYCASTING.BOT_BOT_REVICE_INTERVAL then
 			self._BotReviveBotTimer = 0.0
 			self:_CheckForBotBotRevive()
@@ -1213,7 +1216,7 @@ function BotManager:_CheckForBotBotAttack()
 		local s_BotIdToCheck = self._BotBotAttackList[i]
 		local s_Bot = self:GetBotById(s_BotIdToCheck)
 
-		if s_Bot and s_Bot.m_Player and s_Bot.m_Player.soldier and s_Bot:IsReadyToAttack(false) then
+		if s_Bot and s_Bot.m_Player and s_Bot.m_Player.soldier and s_Bot:IsReadyToAttack(false, nil, false) then
 			local s_BotPosition = nil
 			if s_Bot.m_Player.controlledControllable then
 				s_BotPosition = s_Bot.m_Player.controlledControllable.transform.trans