+++ .github/CHANGELOG.md
@@ -33,6 +33,7 @@ This is the changelog for the version V2.9 Don't forget to [join us on Discord](
 * improve speed on save of traces (by tanger-noisu)
 * improve general behavior
 * fix modlist.realoadextension
+* Priority-System on attack (by tanger-noisu)
 
 ### Bug fixes
 * Merged pull request #323 from tagener-noisu Path-Switcher fixes